// lib/netlify/schema.ts
export interface components {
    schemas: {
// ...
        deploy: {
            id?: string
            site_id?: string
            user_id?: string
            build_id?: string
            state?:
                | 'new'
                | 'pending_review'
                | 'accepted'
                | 'rejected'
                | 'enqueued'
                | 'building'
                | 'uploading'
                | 'uploaded'
                | 'preparing'
                | 'prepared'
                | 'processing'
                | 'processed'
                | 'ready'
                | 'error'
                | 'retrying'
            name?: string
            url?: string
            ssl_url?: string
            admin_url?: string
            deploy_url?: string
            deploy_ssl_url?: string
            screenshot_url?: string
            review_id?: number
            draft?: boolean
            required?: string[]
            required_functions?: string[]
            error_message?: string
            branch?: string
            commit_ref?: string
            commit_url?: string
            skipped?: boolean
            /** Format: dateTime */
            created_at?: string
            /** Format: dateTime */
            updated_at?: string
            /** Format: dateTime */
            published_at?: string
            title?: string
            context?: string
            locked?: boolean
            review_url?: string
            framework?: string
            function_schedules?: components['schemas']['functionSchedule'][]
            commit_message?: string
            deploy_time?: number
            available_functions?: any[]
            committer?: string
            skipped_log?: string
            manual_deploy?: boolean
            plugin_state?: string
            lighthouse_plugin_scores?: Record<string, any>
            links: {
                permalink: string
                alias: string
                branch: null
            }
            entry_path?: string
            views_count?: any
            public_repo?: string
            pending_review_reason?: any
            lighthouse?: any
            edge_functions_present?: any
            expires_at?: string
            blobs_region?: string
            summary?: {
                status: 'ready' | string
                messages: {
                    type: 'info' | string
                    title: string
                    description: string
                    details: null | string
                }[]
            }
        }
// ...
    }
// ...
}